+++ tools/search/filter_test.go
@@ -40,7 +40,8 @@ func TestFilterDataBuildExpr(t *testing.T) {
 		},
 		{
 			"simple expression",
-			"test1 > 1", false,
+			"test1 > 1",
+			false,
 			"^" +
 				regexp.QuoteMeta("[[test1]] > {:") +
 				".+" +
@@ -49,14 +50,16 @@ func TestFilterDataBuildExpr(t *testing.T) {
 		},
 		{
 			"like with 2 columns",
-			"test1 ~ test2", false,
+			"test1 ~ test2",
+			false,
 			"^" +
 				regexp.QuoteMeta("[[test1]] LIKE ('%' || [[test2]] || '%') ESCAPE '\\'") +
 				"$",
 		},
 		{
 			"like with right column operand",
-			"'lorem' ~ test1", false,
+			"'lorem' ~ test1",
+			false,
 			"^" +
 				regexp.QuoteMeta("{:") +
 				".+" +
@@ -65,7 +68,8 @@ func TestFilterDataBuildExpr(t *testing.T) {
 		},
 		{
 			"like with left column operand and text as right operand",
-			"test1 ~ 'lorem'", false,
+			"test1 ~ 'lorem'",
+			false,
 			"^" +
 				regexp.QuoteMeta("[[test1]] LIKE {:") +
 				".+" +
@@ -74,14 +78,16 @@ func TestFilterDataBuildExpr(t *testing.T) {
 		},
 		{
 			"not like with 2 columns",
-			"test1 !~ test2", false,
+			"test1 !~ test2",
+			false,
 			"^" +
 				regexp.QuoteMeta("[[test1]] NOT LIKE ('%' || [[test2]] || '%') ESCAPE '\\'") +
 				"$",
 		},
 		{
 			"not like with right column operand",
-			"'lorem' !~ test1", false,
+			"'lorem' !~ test1",
+			false,
 			"^" +
 				regexp.QuoteMeta("{:") +
 				".+" +
@@ -90,7 +96,8 @@ func TestFilterDataBuildExpr(t *testing.T) {
 		},
 		{
 			"like with left column operand and text as right operand",
-			"test1 !~ 'lorem'", false,
+			"test1 !~ 'lorem'",
+			false,
 			"^" +
 				regexp.QuoteMeta("[[test1]] NOT LIKE {:") +
 				".+" +
@@ -99,7 +106,8 @@ func TestFilterDataBuildExpr(t *testing.T) {
 		},
 		{
 			"current datetime constant",
-			"test1 > @now", false,
+			"test1 > @now",
+			false,
 			"^" +
 				regexp.QuoteMeta("[[test1]] > {:") +
 				".+" +
@@ -113,25 +121,25 @@ func TestFilterDataBuildExpr(t *testing.T) {
 			"^" +
 				regexp.QuoteMeta("(([[test1]] > {:") +
 				".+" +
-				regexp.QuoteMeta("} OR COALESCE([[test2]], '') != COALESCE({:") +
+				regexp.QuoteMeta("} OR [[test2]] != {:") +
 				".+" +
-				regexp.QuoteMeta("}, '')) AND [[test3]] LIKE {:") +
+				regexp.QuoteMeta("}) AND [[test3]] LIKE {:") +
 				".+" +
-				regexp.QuoteMeta("} ESCAPE '\\' AND COALESCE([[test4.sub]], '') = COALESCE(NULL, ''))") +
+				regexp.QuoteMeta("} ESCAPE '\\' AND [[test4.sub]] = '')") +
 				"$",
 		},
 		{
 			"combination of special literals (null, true, false)",
 			"test1=true && test2 != false && test3 = null || test4.sub != null",
 			false,
-			"^" + regexp.QuoteMeta("(COALESCE([[test1]], '') = COALESCE(1, '') AND COALESCE([[test2]], '') != COALESCE(0, '') AND COALESCE([[test3]], '') = COALESCE(NULL, '') OR COALESCE([[test4.sub]], '') != COALESCE(NULL, ''))") + "$",
+			"^" + regexp.QuoteMeta("([[test1]] = 1 AND [[test2]] != 0 AND [[test3]] = '' OR [[test4.sub]] != '')") + "$",
 		},
 		{
 			"all operators",
 			"(test1 = test2 || test2 != test3) && (test2 ~ 'example' || test2 !~ '%%abc') && 'switch1%%' ~ test1 && 'switch2' !~ test2 && test3 > 1 && test3 >= 0 && test3 <= 4 && 2 < 5",
 			false,
 			"^" +
-				regexp.QuoteMeta("((COALESCE([[test1]], '') = COALESCE([[test2]], '') OR COALESCE([[test2]], '') != COALESCE([[test3]], '')) AND ([[test2]] LIKE {:") +
+				regexp.QuoteMeta("(([[test1]] = [[test2]] OR [[test2]] != [[test3]]) AND ([[test2]] LIKE {:") +
 				".+" +
 				regexp.QuoteMeta("} ESCAPE '\\' OR [[test2]] NOT LIKE {:") +
 				".+" +